rename view-correction module to correction
