+++ src/pages/corrections/index.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next'
 
-import CorrectionsList from '../../modules/view-correction/components/CorrectionsList'
+import CorrectionsList from '../../modules/correction/components/CorrectionsList'
 
 const CorrectionsPage: NextPage = () => <CorrectionsList />
